Move burnable token behavior file to behaviors dir
